feat: Refactor styles and components for SignupTab and InstructionsTab; add new BookTab and ExampleTab styles
