--- app/components/aboutPolydotsTab/IntroTab.module.css
@@ -1,35 +1,0 @@
-.titleBox {
-	position: absolute;
-	top: 0;
-	width: 100%;
-	height: 200px;
-	background-color: var(--myrtle-green);
-	border-bottom: 1px solid var(--pale-silver);
-
-	color: white;
-	display: flex;
-	flex-direction: column;
-	align-items: center;
-	justify-content: space-evenly;
-}
-
-.titleBox h1 {
-	font-size: var(--font-size-xxl);
-	margin-bottom: var(--spacing-sm);
-	font-family: "Consolas", monospace;
-}
-
-.titleBox h2 {
-	font-size: var(--font-size-xl);
-	font-family: "Consolas", monospace;
-	font-weight: 300;
-}
-
-.infoBox {
-	display: flex;
-	flex-direction: column;
-	height: calc(100% - 200px);
-	position: absolute;
-	top: 200px;
-	padding: var(--spacing-xl);
-}
