--- app/components/exampleTab/ExampleTab.module.css
@@ -1,24 +1,0 @@
-.title {
-	font-family: "Garamond", serif;
-	color: var(--moon-white);
-	font-size: var(--font-size-xxxl);
-}
-
-.bookCoverContainer {
-	margin-top: var(--spacing-lg);
-	max-width: 100%;
-	width: 100%;
-}
-
-.bookCover {
-	object-fit: cover;
-}
-
-.description {
-	font-family: "Garamond", serif;
-	font-size: var(--font-size-xxl);
-	font-style: italic;
-	font-weight: lighter;
-	color: var(--moon-white);
-	text-align: center;
-}
