+++ app/tabs/InstructionsTab/Buttons/BrushIcon.tsx
