+++ app/tabs/InstructionsTab/InstructionsTab.module.css
