+++ app/tabs/InstructionsTab/Letters/Letters.module.css
@@ -12,11 +12,36 @@
 	margin: var(--spacing-xxs);
 }
 
+@media (max-width: 600px) {
+	.letters {
+		font-size: var(--font-size-md);
+	}
+}
+
+@media (max-width: 400px) {
+	.letters {
+		font-size: var(--font-size-sm);
+	}
+}
+
 .letterCurrent {
 	color: var(--saturated-earth-yellow);
 }
 
 .letterPrevious {
+	color: var(--saturated-myrtle-green);
+}
+
+.letterNext {
+	color: var(--saturated-mountbatten-pink);
+	animation: scaleUp 1s ease-in-out forwards;
+}
+
+.letterFuture {
+	color: var(--pale-silver);
+}
+
+.letterPast {
 	color: var(--saturated-myrtle-green);
 }
 
@@ -34,21 +59,11 @@
 	}
 }
 
-.letterNext {
-	color: var(--saturated-mountbatten-pink);
-	animation: scaleUp 1s ease-in-out forwards;
-}
-
-.letterFuture {
-	color: var(--pale-silver);
-}
-
-.letterPast {
-	color: var(--saturated-myrtle-green);
-}
-
 .colorIndicator {
-	margin-left: var(--spacing-md);
+	text-transform: uppercase;
+	font-family: sans-serif;
+	font-size: var(--font-size-xs);
+	color: var(--pale-silver);
 }
 
 .timeToColor {
